add DirTree
