+++ components/tree/style/index.less
@@ -2,6 +2,7 @@
 @import "../../style/mixins/index";
 @import "../../checkbox/style/mixin";
 @import "./mixin";
+@import "./directory";
 
 @tree-prefix-cls: ~"@{ant-prefix}-tree";
 @tree-showline-icon-color: @text-color-secondary;
