+++ package.json
@@ -73,7 +73,7 @@
     "rc-tabs": "~9.2.0",
     "rc-time-picker": "~3.3.0",
     "rc-tooltip": "~3.7.0",
-    "rc-tree": "~1.8.0",
+    "rc-tree": "~1.9.0",
     "rc-tree-select": "~1.12.0",
     "rc-upload": "~2.4.0",
     "rc-util": "^4.0.4",
